requirements.txt
python-telegram-bot>=21.3
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2
